+++ Project/src/BusinessLayer/Shops/Basket.java
@@ -1,0 +1,11 @@
+package BusinessLayer.Shops;
+
+import java.util.Collection;
+
+public class Basket {
+
+    private int id;
+    private Shop shop;
+    private Collection<Product> products;
+
+}
