+++ Project/src/BusinessLayer/Shops/DiscountPolicy.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops;
+
+public abstract class DiscountPolicy {
+}
