+++ Project/src/BusinessLayer/Shops/Product.java
@@ -1,0 +1,9 @@
+package BusinessLayer.Shops;
+
+interface  Product {
+    int getID();
+    String getName();
+    double getPrice();
+    int getQuantity();
+
+}
