+++ Project/src/BusinessLayer/Shops/Purchase.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops;
+
+public class Purchase {
+}
