+++ Project/src/BusinessLayer/Shops/PurchaseHistory.java
@@ -1,0 +1,9 @@
+package BusinessLayer.Shops;
+
+import java.util.Collection;
+
+public class PurchaseHistory {
+
+    Collection<Purchase> past_purchases;
+
+}
