+++ Project/src/BusinessLayer/Shops/PurchaseHistoryServices.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops;
+
+public class PurchaseHistoryServices {
+}
