+++ Project/src/BusinessLayer/Shops/PurchasePolicy.java
@@ -1,0 +1,5 @@
+package BusinessLayer.Shops;
+
+public abstract class PurchasePolicy {
+
+}
