+++ Project/src/BusinessLayer/Shops/Shop.java
@@ -1,0 +1,13 @@
+package BusinessLayer.Shops;
+
+
+public class Shop {
+    public enum State {
+        OPEN,
+        CLOSED
+    }
+    private int id;
+    private String name;
+    private State state;
+
+}
