+++ Project/src/BusinessLayer/Shops/System/Notification.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.System;
+
+public class Notification {
+}
