+++ Project/src/BusinessLayer/Shops/System/Payment.java
@@ -1,0 +1,5 @@
+package BusinessLayer.Shops.System;
+
+public class Payment {
+    int id;
+}
