+++ Project/src/BusinessLayer/Shops/System/Supply.java
@@ -1,0 +1,5 @@
+package BusinessLayer.Shops.System;
+
+public class Supply {
+    int id;
+}
