+++ Project/src/BusinessLayer/Shops/Users/AssignShopManager.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class AssignShopManager extends BaseAction {
+}
