+++ Project/src/BusinessLayer/Shops/Users/AssignShopOwner.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class AssignShopOwner extends BaseAction{
+}
