+++ Project/src/BusinessLayer/Shops/Users/BaseAction.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public abstract class BaseAction{
+}
