+++ Project/src/BusinessLayer/Shops/Users/ChangeManagerPermission.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class ChangeManagerPermission extends BaseAction{
+}
