+++ Project/src/BusinessLayer/Shops/Users/CloseShop.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class CloseShop extends BaseAction{
+}
