+++ Project/src/BusinessLayer/Shops/Users/DiscountAndShoppingPolicy.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class DiscountAndShoppingPolicy extends BaseAction{
+}
