+++ Project/src/BusinessLayer/Shops/Users/Guest.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class Guest extends User {
+}
