+++ Project/src/BusinessLayer/Shops/Users/HistoryInfo.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class HistoryInfo extends BaseAction{
+}
