+++ Project/src/BusinessLayer/Shops/Users/RolesInfo.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class RolesInfo extends BaseAction{
+}
