+++ Project/src/BusinessLayer/Shops/Users/ShopAdministrator.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class ShopAdministrator extends SubscribedUser{
+}
