+++ Project/src/BusinessLayer/Shops/Users/ShopManager.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class ShopManager extends ShopAdministrator{
+}
