+++ Project/src/BusinessLayer/Shops/Users/ShopOwner.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class ShopOwner extends ShopAdministrator{
+}
