+++ Project/src/BusinessLayer/Shops/Users/SubscribedUser.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class SubscribedUser extends User {
+}
