+++ Project/src/BusinessLayer/Shops/Users/SupplyManagment.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class SupplyManagment extends BaseAction {
+}
