+++ Project/src/BusinessLayer/Shops/Users/SystemManager.java
@@ -1,0 +1,4 @@
+package BusinessLayer.Shops.Users;
+
+public class SystemManager extends SubscribedUser {
+}
