+++ Project/src/BusinessLayer/Shops/Users/User.java
@@ -1,0 +1,5 @@
+package BusinessLayer.Shops.Users;
+
+public abstract class User {
+
+}
