--- Project/src/Shops/Basket.java
@@ -1,11 +1,0 @@
-package Shops;
-
-import java.util.Collection;
-
-public class Basket {
-
-    private int id;
-    private Shop shop;
-    private Collection<Product> products;
-
-}
