--- Project/src/Shops/DiscountPolicy.java
@@ -1,4 +1,0 @@
-package Shops;
-
-public abstract class DiscountPolicy {
-}
