--- Project/src/Shops/Product.java
@@ -1,9 +1,0 @@
-package Shops;
-
-interface  Product {
-    int getID();
-    String getName();
-    double getPrice();
-    int getQuantity();
-
-}
