--- Project/src/Shops/Purchase.java
@@ -1,4 +1,0 @@
-package Shops;
-
-public class Purchase {
-}
