--- Project/src/Shops/PurchaseHistory.java
@@ -1,9 +1,0 @@
-package Shops;
-
-import java.util.Collection;
-
-public class PurchaseHistory {
-
-    Collection<Purchase> past_purchases;
-
-}
