--- Project/src/Shops/PurchaseHistoryServices.java
@@ -1,4 +1,0 @@
-package Shops;
-
-public class PurchaseHistoryServices {
-}
