--- Project/src/Shops/PurchasePolicy.java
@@ -1,5 +1,0 @@
-package Shops;
-
-public abstract class PurchasePolicy {
-
-}
