--- Project/src/Shops/Shop.java
@@ -1,13 +1,0 @@
-package Shops;
-
-
-public class Shop {
-    public enum State {
-        OPEN,
-        CLOSED
-    }
-    private int id;
-    private String name;
-    private State state;
-
-}
