--- Project/src/System/Notification.java
@@ -1,4 +1,0 @@
-package System;
-
-public class Notification {
-}
