--- Project/src/System/Payment.java
@@ -1,5 +1,0 @@
-package System;
-
-public class Payment {
-    int id;
-}
