--- Project/src/System/Supply.java
@@ -1,5 +1,0 @@
-package System;
-
-public class Supply {
-    int id;
-}
