--- Project/src/Users/AssignShopManager.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class AssignShopManager extends BaseAction {
-}
