--- Project/src/Users/AssignShopOwner.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class AssignShopOwner extends BaseAction{
-}
