--- Project/src/Users/BaseAction.java
@@ -1,4 +1,0 @@
-package Users;
-
-public abstract class BaseAction{
-}
