--- Project/src/Users/ChangeManagerPermission.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class ChangeManagerPermission extends BaseAction{
-}
