--- Project/src/Users/CloseShop.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class CloseShop extends BaseAction{
-}
