--- Project/src/Users/DiscountAndShoppingPolicy.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class DiscountAndShoppingPolicy extends BaseAction{
-}
