--- Project/src/Users/Guest.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class Guest extends User {
-}
