--- Project/src/Users/HistoryInfo.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class HistoryInfo extends BaseAction{
-}
