--- Project/src/Users/RolesInfo.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class RolesInfo extends BaseAction{
-}
