--- Project/src/Users/ShopAdministrator.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class ShopAdministrator extends SubscribedUser{
-}
