--- Project/src/Users/ShopManager.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class ShopManager extends ShopAdministrator{
-}
