--- Project/src/Users/ShopOwner.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class ShopOwner extends ShopAdministrator{
-}
