--- Project/src/Users/SubscribedUser.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class SubscribedUser extends User {
-}
