--- Project/src/Users/SupplyManagment.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class SupplyManagment extends BaseAction {
-}
