--- Project/src/Users/SystemManager.java
@@ -1,4 +1,0 @@
-package Users;
-
-public class SystemManager extends SubscribedUser {
-}
