--- Project/src/Users/User.java
@@ -1,5 +1,0 @@
-package Users;
-
-public abstract class User {
-
-}
